Fix buffer overflow: If len == size, then NUL would get written one past the end of s.
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@3331 8aca7d54-2c30-db11-9de9-000461428c89

+++ dataobj/loadsave.h
@@ -102,7 +102,7 @@ public:
 	void rdwr_str(const char *&s);
 
 	// s is a buf of size given
-	void rdwr_str(char *s, int size);
+	void rdwr_str(char* s, size_t size);
 
 	// only meaningful for XML
 	void start_tag( const char *tag );
